jetls-client: Move TypeScript sources under src
The extension implementation files were mixed with package metadata
and build scripts at the `jetls-client` root.

Move the production TypeScript sources into `jetls-client/src` and
update the esbuild entrypoint and Node test imports. The extension
bundle remains at `out/jetls-client.js`.
